Move OasAgent::Version to the correct place
