+++ lib/oas_agent.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 # frozen_string_literal: true
 
-require "agent/version"
+require "oas_agent/version"
 require "oas_agent/control"
 
 if defined?(Rails::VERSION)
