+++ oas_agent.gemspec
@@ -1,6 +1,6 @@
 lib = File.expand_path("lib", File.dirname(__FILE__))
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require "agent/version"
+require "oas_agent/version"
 
 Gem::Specification.new do |spec|
   spec.name          = "oas_agent"
